Next

+++ Makefile
@@ -23,17 +23,6 @@ CLONE_DIR  = clone
 REPOSITORY = git@github.com:dunglas/symfony-docker.git
 
 #
-# OVERLOADING
-#
-
--include overload/.env
-
-PROJECT_NAME           ?= $(shell basename $(CURDIR))
-COMPOSE_BUILD_OPTS     ?=
-COMPOSE_UP_SERVER_NAME ?= $(PROJECT_NAME).localhost
-COMPOSE_UP_ENV_VARS    ?=
-
-#
 # SYMFONY ENVIRONMENT VARIABLES
 #
 
@@ -57,9 +46,9 @@ $(info Warning: APP_ENV is overloaded outside .env and .env.local files)
 endif
 
 ifeq ($(FILE_ENV),prod)
-$(info Warning: Your are in the prod environment)
+$(info Warning: Your are in the PROD environment)
 else ifeq ($(FILE_ENV),test)
-$(info Warning: Your are in the test environment)
+$(info Warning: Your are in the TEST environment)
 endif
 
 # @see https://symfony.com/doc/current/deployment.html#b-configure-your-environment-variables
@@ -78,21 +67,47 @@ ifndef COMPOSE_V2
 $(error Docker Compose CLI plugin is required but is not available on your system)
 endif
 
-COMPOSE_BASE     = compose.yaml
-COMPOSE_OVERRIDE = compose.override.yaml
-COMPOSE_PROD     = compose.prod.yaml
-COMPOSE_PREFIX   = docker compose -p $(PROJECT_NAME) -f $(COMPOSE_BASE)
+COMPOSE = docker compose -p $(PROJECT_NAME)
 
 ifeq ($(FILE_ENV),prod)
-COMPOSE = $(COMPOSE_PREFIX) -f $(COMPOSE_PROD)
-else
-COMPOSE = $(COMPOSE_PREFIX) -f $(COMPOSE_OVERRIDE)
+COMPOSE = $(COMPOSE) -f compose.yaml -f compose.prod.yaml
 endif
 
 CONTAINER_PHP = $(COMPOSE) exec php
 PHP           = $(CONTAINER_PHP) php
 COMPOSER      = $(CONTAINER_PHP) composer
 CONSOLE       = $(PHP) bin/console
+
+#
+# OPTIONS
+#
+
+-include .env.options.local
+
+# https://symfony.com/releases
+SYMFONY_LTS     = 6.4.*
+SYMFONY_VERSION ?= $(SYMFONY_LTS)
+BUILD_ENV       ?= SYMFONY_VERSION=$(SYMFONY_VERSION)
+
+PROJECT_NAME    ?= $(shell basename $(CURDIR))
+SERVER_NAME     ?= $(PROJECT_NAME).localhost
+UP_ENV          ?= SERVER_NAME=$(SERVER_NAME)
+
+ifneq ($(XDEBUG_MODE),)
+UP_ENV = $(UP_ENV) XDEBUG_MODE=$(XDEBUG_MODE)
+endif
+
+ifneq ($(HTTP_PORT),)
+UP_ENV = $(UP_ENV) HTTP_PORT=$(HTTP_PORT)
+endif
+
+ifneq ($(HTTPS_PORT),)
+UP_ENV = $(UP_ENV) HTTPS_PORT=$(HTTPS_PORT)
+endif
+
+ifneq ($(HTTPS_PORT),)
+UP_ENV = $(UP_ENV) HTTPS_PORT=$(HTTPS_PORT)
+endif
 
 ## — 🐳 🎵 THE SYMFONY STARTER MAKEFILE 🎵 🐳 —————————————————————————————————
 
@@ -103,35 +118,26 @@ CONSOLE       = $(PHP) bin/console
 .DEFAULT_GOAL = help
 .PHONY: help
 help: ## Print self-documented Makefile
-	@grep -E '(^[.a-zA-Z_-]+[^:]+:.*##.*?$$)|(^#{2})' Makefile \
-	| awk 'BEGIN {FS = "## "}; \
-		{ \
-			split($$1, line, ":"); \
-			targets=line[1]; \
-			description=$$2; \
-			if (targets == "##") { \
-				# --- space --- \
-				printf "\033[33m%s\n", ""; \
-			} else if (targets == "" && description != "") { \
-				# --- title --- \
-				printf "\033[33m\n%s\n", description; \
-			} else if (targets != "" && description != "") { \
-				# --- target, alias, description --- \
-				split(targets, parts, " "); \
-				target=parts[1]; \
-				alias=parts[2]; \
-				printf "\033[32m  %-26s \033[34m%-2s \033[0m%s\n", target, alias, description; \
-			} \
-		}'
+	@grep -E '(^[.a-zA-Z_-]+[^:]+:.*##.*?$$)|(^#{2})' Makefile | awk 'BEGIN {FS = "## "}; { \
+		split($$1, line, ":"); targets=line[1]; description=$$2; \
+		if (targets == "##") { \
+			printf "\033[33m%s\n", ""; # space \
+		} else if (targets == "" && description != "") { \
+			printf "\033[33m\n%s\n", description; # title \
+		} else if (targets != "" && description != "") { \
+			split(targets, parts, " "); target=parts[1]; alias=parts[2]; \
+			printf "\033[32m  %-26s \033[34m%-2s \033[0m%s\n", target, alias, description; # target alias: description \
+		} \
+	}'
 	@echo
 
 ## — PROJECT 🚀 ———————————————————————————————————————————————————————————————
 
 .PHONY: generate
-generate: confirm_continue clone build up_d permissions info ## Generate a fresh Symfony application with the Docker configuration
+generate: clone build up permissions info ## Generate a fresh Symfony application with the Docker configuration
 
 .PHONY: start
-start: up_d info ## Start the project (implies detached mode)
+start: up info ## Start the project (implies detached mode)
 
 .PHONY: stop
 stop: down ## Stop the project
@@ -139,33 +145,27 @@ stop: down ## Stop the project
 .PHONY: restart
 restart: stop start ## Restart the project
 
-##
-
 .PHONY: clean
-clean: confirm_continue ## Remove all generated files [y/N]
-	rm -rf $(CLONE_DIR) \
-      .github bin config docs frankenphp public src var vendor
+clean: ## Remove all generated files and restore original files
+	git restore LICENSE
+	rm -rf $(CLONE_DIR).github bin config docs frankenphp public src var vendor
 	rm  -f \
-      .dockerignore .editorconfig .env .gitattributes .gitignore \
+      .dockerignore .editorconfig .env .env.dev .gitattributes .gitignore \
       compose.override.yaml compose.prod.yaml compose.yaml \
       composer.json composer.lock Dockerfile symfony.lock
 
-##
-
 PHONY: info
-info i: ## Show info
-	@$(MAKE) -s overload_file env_files vars
+info: ## Show info
+	@$(MAKE) -s vars
 	@printf "\n$(Y)Info$(S)"
 	@printf "\n$(Y)----$(S)\n\n"
-	@printf "* Go on $(G)https://$(COMPOSE_UP_SERVER_NAME)/$(S)\n"
 	@printf "* Run $(Y)make$(S) to see all shorcuts for the most common tasks.\n"
-	@printf "* Run $(Y). aliases$(S) to load all the project aliases.\n"
+	@printf "* Go on $(G)https://$(SERVER_NAME)/$(S)\n"
 
 ## — SYMFONY 🎵 ———————————————————————————————————————————————————————————————
 
 .PHONY: symfony
 symfony sf: ## Run Symfony - $ make symfony [p=<params>] - Example: $ make symfony p=cache:clear
-	@$(eval p ?=)
 	$(CONSOLE) $(p)
 
 .PHONY: cc
@@ -180,10 +180,6 @@ cw: ## Warm up an empty cache
 about: ## Display information about the current project
 	$(CONSOLE) about
 
-.PHONY: dotenv
-dotenv: ## Lists all dotenv files with variables and values
-	$(CONSOLE) debug:dotenv
-
 .PHONY: dumpenv
 dumpenv: ## Generate .env.local.php (PROD)
 	$(COMPOSER) dump-env prod
@@ -192,44 +188,30 @@ dumpenv: ## Generate .env.local.php (PROD)
 
 .PHONY: php
 php: ## Run PHP - $ make php [p=<params>]- Example: $ make php p=--version
-	@$(eval p ?=)
 	$(PHP) $(p)
 
 php_sh: ## Connect to the PHP container
 	$(CONTAINER_PHP) sh
 
-php_version: ## PHP version number
-	$(PHP) -v
-
-php_modules: ## Show compiled in modules
-	$(PHP) -m
-
 ## — COMPOSER 🧙 ——————————————————————————————————————————————————————————————
 
 .PHONY: composer
 composer: ## Run composer - $ make composer [p=<params>] - Example: $ make composer p="require --dev phpunit/phpunit"
-	@$(eval p ?=)
 	$(COMPOSER) $(p)
 
-composer_version: ## Composer version
-	$(COMPOSER) --version
-
-composer_validate: ## Validate composer.json and composer.lock
-	$(COMPOSER) validate --strict --check-lock
-
-##
-
 composer_install: ## Install packages using composer
-	$(COMPOSER) install
-
-composer_install@prod: ## Install packages using composer (PROD)
+ifeq ($(FILE_ENV),prod)
 	$(COMPOSER) install --verbose --prefer-dist --no-progress --no-interaction --no-dev --optimize-autoloader
+else
+	$(COMPOSER) install
+endif
 
 composer_update: ## Update packages using composer
-	$(COMPOSER) update
-
-composer_update@prod: ## Update packages using composer (PROD)
+ifeq ($(FILE_ENV),prod)
 	$(COMPOSER) update --verbose --prefer-dist --no-progress --no-interaction --no-dev --optimize-autoloader
+else
+	$(COMPOSER) update
+endif
 
 ## — SYMFONY DOCKER 🎵 🐳 —————————————————————————————————————————————————————
 
@@ -255,12 +237,8 @@ endif
 ## — DOCKER 🐳 ————————————————————————————————————————————————————————————————
 
 .PHONY: up
-up: ## Start the container - $ make up [p=<params>] - Example: $ make up p=-d
-	@$(eval p ?=)
-	SERVER_NAME=$(COMPOSE_UP_SERVER_NAME) $(COMPOSE_UP_ENV_VARS) $(COMPOSE) up --remove-orphans --pull always $(p)
-
-up_d: ## Start the container (wait for services to be running|healthy - detached mode)
-	$(MAKE) up p="--wait -d"
+up: ## Start the container - $ make up [p=<params>] - Example: $ make up p=-d (wait for services to be running|healthy - detached mode by default)
+	$(UP_ENV) $(COMPOSE) up --remove-orphans --pull always --wait $(p)
 
 .PHONY: down
 down: ## Stop the container
@@ -268,66 +246,26 @@ down: ## Stop the container
 
 .PHONY: build
 build: ## Build or rebuild services - $ make build [p=<params>] - Example: $ make build p=--no-cache
-	@$(eval p ?=)
-	$(COMPOSE) build $(COMPOSE_BUILD_OPTS) $(p)
+	$(BUILD_ENV) $(COMPOSE) build $(p)
 
 .PHONY: logs
 logs: ## See the container’s logs
 	$(COMPOSE) logs -f
 
-##
-
-docker_stop_all: confirm_continue ## Stop all running containers [y/N]
-	docker stop $$(docker ps -a -q)
-
-docker_remove_all: confirm_continue ## Remove all stopped containers [y/N]
-	docker rm $$(docker ps -a -q)
-
 ## — TROUBLESHOOTING 😵‍️ ———————————————————————————————————————————————————————
 
 .PHONY: permissions
 permissions p: ## Run it if you cannot edit some of the project files on Linux (https://github.com/dunglas/symfony-docker/blob/main/docs/troubleshooting.md)
-	@printf "\n$(Y)Permissions$(S)"
-	@printf "\n$(Y)-----------$(S)\n\n"
+	@printf "\n$(Y)Permissions (Linux)$(S)"
+	@printf "\n$(Y)-------------------$(S)\n\n"
+ifeq ($(PERMISSIONS),on)
 	$(COMPOSE) run --rm php chown -R $(USER_ID):$(GROUP_ID) .
 	@printf " $(G)✔$(S) You are now defined as the owner $(Y)$(USER_ID):$(GROUP_ID)$(S) of the project files.\n"
+else
+	@printf " $(G)✔$(S) Nothing to do (add PERMISSIONS=on to .env.options.local to activate `permissions` command).\n"
+endif
 
 ## — UTILS 🛠️  —————————————————————————————————————————————————————————————————
-
-overload_file: ## Show overload file loaded into that Makefile
-	@printf "\n$(Y)Overload file$(S)"
-	@printf "\n$(Y)-------------$(S)\n\n"
-	@printf "File loaded into that Makefile:\n\n"
-ifneq ("$(wildcard overload/.env)","")
-	@printf "* $(G)✔$(S) overload/.env\n"
-else
-	@printf "* $(R)⨯$(S) overload/.env\n"
-endif
-
-env_files: ## Show Symfony env files loaded into that Makefile
-	@printf "\n$(Y)Symfony env files$(S)"
-	@printf "\n$(Y)-----------------$(S)\n\n"
-	@printf "Files loaded into that Makefile (in order of decreasing priority) $(Y)[FILE_ENV=$(FILE_ENV)]$(S):\n\n"
-ifneq ("$(wildcard .env.$(FILE_ENV).local)","")
-	@printf "* $(G)✔$(S) .env.$(FILE_ENV).local\n"
-else
-	@printf "* $(R)⨯$(S) .env.$(FILE_ENV).local\n"
-endif
-ifneq ("$(wildcard .env.$(FILE_ENV))","")
-	@printf "* $(G)✔$(S) .env.$(FILE_ENV)\n"
-else
-	@printf "* $(R)⨯$(S) .env.$(FILE_ENV)\n"
-endif
-ifneq ("$(wildcard .env.local)","")
-	@printf "* $(G)✔$(S) .env.local\n"
-else
-	@printf "* $(R)⨯$(S) .env.local\n"
-endif
-ifneq ("$(wildcard .env)","")
-	@printf "* $(G)✔$(S) .env\n"
-else
-	@printf "* $(R)⨯$(S) .env\n"
-endif
 
 .PHONY: vars
 vars: ## Show variables
@@ -339,39 +277,14 @@ vars: ## Show variables
 	@printf "\n$(G)BASE$(S)\n"
 	@printf "  CLONE_DIR : $(CLONE_DIR)\n"
 	@printf "  REPOSITORY: $(REPOSITORY)\n"
-	@printf "\n$(G)OVERLOADING$(S)\n"
-	@printf "  PROJECT_NAME          : $(PROJECT_NAME)\n"
-	@printf "  COMPOSE_BUILD_OPTS    : $(COMPOSE_BUILD_OPTS)\n"
-	@printf "  COMPOSE_UP_SERVER_NAME: $(COMPOSE_UP_SERVER_NAME)\n"
-	@printf "  COMPOSE_UP_ENV_VARS   : $(COMPOSE_UP_ENV_VARS)\n"
+	@printf "\n$(G)OPTIONS$(S)\n"
+	@printf "  PERMISSIONS : $(PERMISSIONS)\n"
+	@printf "  PROJECT_NAME: $(PROJECT_NAME)\n"
+	@printf "  BUILD_ENV   : $(BUILD_ENV)\n"
+	@printf "  UP_ENV      : $(UP_ENV)\n"
 	@printf "\n$(G)SYMFONY ENVIRONMENT VARIABLES$(S)\n"
 	@printf "  APP_ENV   : $(APP_ENV)\n"
 	@printf "  APP_SECRET: $(APP_SECRET)\n"
 	@printf "\n$(G)DOCKER$(S)\n"
 	@printf "  COMPOSE_V2: $(COMPOSE_V2)\n"
 	@printf "  COMPOSE   : $(COMPOSE)\n"
-
-## — INTERNAL 🚧‍️ ——————————————————————————————————————————————————————————————
-
-.PHONY: confirm
-confirm: ## Display a confirmation before executing a makefile command - @$(MAKE) -s confirm question=<question> [make_yes=<command>] [make_no=<command>] [no_interaction=<bool>]
-	@$(if $(question),, $(error question argument is required))   # Question to display
-	@$(eval make_yes ?=)                                          # Makefile commands to execute on yes
-	@$(eval make_no ?=)                                           # Makefile commands to execute on no
-	@$(eval no_interaction ?=)                                    # Interactive question or not
-	@\
-	question=$${question:-"Confirm?"}; \
-	if [ "$${no_interaction}" != "true" ]; then \
-		printf "$(G)$${question}$(S) [$(Y)y/N$(S)]: " && read answer; \
-	fi; \
-	answer=$${answer:-N}; \
-	if [ "$${answer}" = y ] || [ "$${answer}" = Y ] || [ "$${no_interaction}" = "true" ]; then \
-		[ -z "$$make_yes" ] && printf "$(Y)(YES) no action!$(S)\n" || $(MAKE) -s $$make_yes no_interaction=true; \
-	else \
-		[ -z "$$make_no" ] && printf "$(Y)(NO) no action!$(S)\n" || $(MAKE) -s $$make_no; \
-	fi
-
-confirm_continue: ## Display a confirmation before continuing [y/N]
-	@$(eval no_interaction ?=) # Interactive question or not
-	@if [ "$${no_interaction}" = "true" ]; then exit 0; fi; \
-	printf "$(G)Do you want to continue?$(S) [$(Y)y/N$(S)]: " && read answer && [ $${answer:-N} = y ]
